Se cambio setBackground para que reciba 3 floats en lugar de 3 ints. Se agregó el load de xml (incompleto)

+++ trunk/src/project/common/ISceneHandler.h
@@ -6,7 +6,7 @@
 class ISceneHandler
 {
     public:
-        virtual void setBackground(int, int, int) =0;
+        virtual void setBackground(float, float, float) =0;
 
         virtual void addCamera(string id) =0;
         virtual void setCameraPos(string camId, float x, float y, float z) =0;
